requirements.txt
altair==5.4.1
altgraph==0.17.4
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
extra-streamlit-components==0.1.71
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.11.1
numpy==2.1.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzbar==0.1.9
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
streamlit==1.39.0
streamlit-authenticator==0.4.1
streamlit-calendar==1.2.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
ttkbootstrap==1.10.1
ttkthemes==3.2.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
XlsxWriter==3.2.0
